Add files via upload
The scatter plot of the Grubbs test is mainly used to clean whether the data has extreme values. The scatter plot can also be used to analyze whether the data is distributed stably, which is used to check the data distribution of the questionnaire or database in the early stage to prevent invalid data.