Add files via upload
GP Neural Network Applied to Marketing Model. This code implements a regression model based on back propagation (BP) neural network to solve a multi-input, multi-output prediction problem. Specifically, the task of the code is to predict two output variables (output1, output2) given input variables (input1, input2, input3, input4, input5).

Code Task:

Input Data Processing:

The code reads an Excel file containing inputs and outputs and divides it into training and test sets. The input has 5 features and the output has 2 variables.
The data is standardized (using StandardScaler) to ensure that the input variables of the model are on a similar scale, which helps to speed up the training of the neural network.
Neural Network Model Definition:

The model is a multi-layer fully connected neural network (BPNet) with 5 nodes in the input layer corresponding to 5 input features. There are two hidden layers in the middle, with 256 and 128 neurons respectively, and the activation function uses ReLU. Finally, there is an output layer with 2 output nodes, predicting output1 and output2 respectively.
The model uses Dropout (set to 0.3) to prevent overfitting.
Training process:

Use mean square error (MSE) as the loss function.
Use Adam optimizer and set the learning rate to 0.1.
During the training process, the model will adjust parameters through the BP algorithm and use 16-size batch data for 100 rounds of iterative training.
Test set evaluation:

On the test set, the code evaluates the performance of the model through mean square error (MSE), mean absolute error (MAE), and R² score. MSE is used to measure the average square of the difference between the predicted value and the actual value, MAE represents the average absolute value of the error, and R² is used to evaluate the degree of fit of the model.