Add files via upload
Regression breakpoint is a kind of "quasi-natural experiment" type of research. Its idea is that there is a continuous variable (driving variable X, or grouping variable, or treatment variable), which can be split into the left and right sides at a critical point cutoff, and then study the impact of this variable on another variable (outcome variable Y, or dependent variable). For example, the undergraduate line in the college entrance examination is 500 points, so some students are slightly lower than 500 points, and some are slightly higher than 500 points. However, it is precisely because of the cutoff value of 500 that determines whether a student can go to college, and ultimately affects the student's future income. Here, the college entrance examination score is the driving variable X, and the income is the outcome variable Y.

This code is a simplified version. If you need specific values, you can customize it or refer to this link (need to be translated into English, original text Chinese): https://spssau.com/helps/conometricstudy/rdd.html